+++ Cargo.toml
@@ -1,5 +1,13 @@
 [workspace]
-members = ["dim", "dim-auth", "dim-database", "dim-events", "dim-extern-api"]
+members = [
+    "dim",
+    "dim-auth",
+    "dim-database",
+    "dim-events",
+    "dim-extern-api",
+    "dim-web",
+    "dim-utils"
+]
 
 [workspace.package]
 version = "0.4.0-dev"
