+++ dim-database/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "dim-database"
+build = "build.rs"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+dim-auth = { path = "../dim-auth" }
+
+serde = { version = "^1", features = ["derive"] }
+serde_derive = "^1"
+serde_json = "^1"
+lazy_static = "^1"
+tracing-log = "0.1.2"
+tracing = "0.1.29"
+ring = "^0.16.11"
+uuid = { version = "0.8.1", features = ["v4"] }
+cfg-if = "1.0.0"
+sqlx = { version = "0.5", features = ["runtime-tokio-rustls", "sqlite"] }
+once_cell = "1.8.0"
+tokio = "1.20.4"
+base64 = "0.13.0"
+itertools = "0.10.3"
+thiserror = "1.0.30"
+displaydoc = "0.2.3"
+
+[dev-dependencies]
+tokio = { version = "1", default-features = false, features = ["rt", "macros"] }
+
+[build-dependencies]
+sqlx = { version = "0.5", features = ["runtime-tokio-rustls"] }
+tokio = "1.20.4"
+dotenv = "0.15.0"
