+++ dim-events/src/lib.rs
@@ -1,3 +1,5 @@
+#![deny(warnings)]
+
 use serde::Serialize;
 use std::collections::HashMap;
