+++ dim-web/Cargo.toml
@@ -6,8 +6,9 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-database = { path = "../database", default-features = false, features = ["sqlite"] }
+dim-database = { path = "../dim-database" }
 dim-utils = { path = "../dim-utils" }
+dim-events =  { path = "../dim-events" }
 
 nightfall = { git = "https://github.com/Dusk-Labs/nightfall", tag = "0.3.12-rc4", default-features = false, features = ["cuda", "ssa_transmux"] }
