+++ dim-web/src/routes/mod.rs
@@ -1,1 +1,1 @@
-pub mod websocket {}
+pub mod websocket;
