+++ dim/src/utils.rs
@@ -1,2 +1,2 @@
 pub use dim_utils::*;
-pub use dim_utils::{balanced_or_tree, warp_try, warp_unwrap};
+pub use dim_utils::{balanced_or_tree, json, warp_try, warp_unwrap};
